refactor: ProductRequest DTO 위치 이동
gift.model -> gift.controller
