+++ src/main/java/gift/controller/ProductRequest.java
@@ -1,4 +1,4 @@
-package gift.model;
+package gift.controller;
 
 public record ProductRequest(String name, int price, String imageUrl) {
 }
